% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convertPAMS}
\alias{convertPAMS}
\title{Interprets text and stores it in a data frame.}
\usage{
convertPAMS(lines, dfResults)
}
\arguments{
\item{lines}{The text of the results to convert.}

\item{dfResults}{A prebuilt data frame to store results into}
}
\description{
Interprets text and stores it in a data frame.
}
