% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convertRH}
\alias{convertRH}
\title{Interprets text and stores it in a data frame.}
\usage{
convertRH(lines, dfResults, hasRaceNumbers)
}
\arguments{
\item{lines}{The text of the results to convert.}

\item{dfResults}{A prebuilt data frame to store results into}

\item{hasRaceNumbers}{TRUE/FALSE the results have race numbers as the first column.}
}
\description{
Interprets text and stores it in a data frame.
}
