% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gradeToInt}
\alias{gradeToInt}
\title{Converts year abbreviations ("Sr", etc) to a grade number (12) (Vectorized).}
\usage{
gradeToInt(grades)
}
\arguments{
\item{grades}{The grade abbreviations to convert}
}
\value{
Integer(s) for each grade.
}
\description{
Converts year abbreviations ("Sr", etc) to a grade number (12) (Vectorized).
}
