% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{keepMedian}
\alias{keepMedian}
\title{Returns a boolean vector with the element that is the median as true
(or the closest element to the average).}
\usage{
keepMedian(times, preferLower = TRUE)
}
\arguments{
\item{times}{A numeric vector to find the median index of.}
}
\value{
A logical vector with 1 true value.
}
\description{
Returns a boolean vector with the element that is the median as true
(or the closest element to the average).
}
