% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{milePace}
\alias{milePace}
\title{Calculates the milepace  (Vectorized).}
\usage{
milePace(raceTime, distance = 3.1)
}
\arguments{
\item{raceTime}{The formatted times (ex: "16:26") to convert.}

\item{distance}{The single race distance (default 3.1 miles).}
}
\value{
A formatted string of each milepace.
}
\description{
Calculates the milepace  (Vectorized).
}
