% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pickMedian}
\alias{pickMedian}
\title{Returns the index of the median, or the closest element. Closest is defined as closest to
the average. An empty vector will see an empty vector returned.}
\usage{
pickMedian(times, preferLower = TRUE)
}
\arguments{
\item{times}{A numeric vector to find the median index of.}

\item{preferLower}{If there is a tie in even length vectors which element to keep.}
}
\value{
The index of the median, or the closest element's index.
}
\description{
Returns the index of the median, or the closest element. Closest is defined as closest to
the average. An empty vector will see an empty vector returned.
}
