% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{placeDot}
\alias{placeDot}
\title{Appends word to the prefix "Place."}
\usage{
placeDot(word)
}
\arguments{
\item{word}{A string of the word to append}
}
\value{
The concatenation of "Place." and word.
}
\description{
Appends word to the prefix "Place."
}
