% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{placeDotC}
\alias{placeDotC}
\title{Appends the first character of word to the prefix "Place."}
\usage{
placeDotC(word)
}
\arguments{
\item{word}{A string of which the first character will be used.}
}
\value{
The concatenation of "Place." and word's first character.
}
\description{
Appends the first character of word to the prefix "Place."
}
