% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{removeDistrictNumber}
\alias{removeDistrictNumber}
\title{Modifies the vector to remove any parenthesis numbers (eg "(3)").}
\usage{
removeDistrictNumber(schoolNames)
}
\arguments{
\item{schoolNames}{The strings to removes numbers from.}
}
\description{
Modifies the vector to remove any parenthesis numbers (eg "(3)").
}
