% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scoreDot}
\alias{scoreDot}
\title{Appends word to the prefix "Score."}
\usage{
scoreDot(word)
}
\arguments{
\item{word}{A string of the word to append}
}
\value{
The concatenation of "Score." and word.
}
\description{
Appends word to the prefix "Score."
}
