% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scoreDotC}
\alias{scoreDotC}
\title{Appends the first character of word to the prefix "Score."}
\usage{
scoreDotC(word)
}
\arguments{
\item{word}{A string of which the first character will be used.}
}
\value{
The concatenation of "Score." and word's first character.
}
\description{
Appends the first character of word to the prefix "Score."
}
