% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sumOfFive}
\alias{sumOfFive}
\title{Sums the first 5 elements of the vector.}
\usage{
sumOfFive(places)
}
\arguments{
\item{places}{A vector to sum.}
}
\value{
The sum of the first 5 elements or -1 if less than 5 are present.
}
\description{
Sums the first 5 elements of the vector.
}
