% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{timeDifference}
\alias{timeDifference}
\title{Calculates and formats the difference between two times.}
\usage{
timeDifference(lowerTime, higherTime)
}
\arguments{
\item{lowerTime}{The shorter time.}

\item{higherTime}{The longer time.}
}
\value{
A formatted string of the difference between the two times.
}
\description{
Calculates and formats the difference between two times.
}
