% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{timeDot}
\alias{timeDot}
\title{Appends word to the prefix "Time."}
\usage{
timeDot(word)
}
\arguments{
\item{word}{A string of the word to append}
}
\value{
The concatenation of "Time." and word.
}
\description{
Appends word to the prefix "Time."
}
