% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{timeDotC}
\alias{timeDotC}
\title{Appends the first character of word to the prefix "Time."}
\usage{
timeDotC(word)
}
\arguments{
\item{word}{A string of which the first character will be used.}
}
\value{
The concatenation of "Time." and word's first character.
}
\description{
Appends the first character of word to the prefix "Time."
}
