% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{toTime}
\alias{toTime}
\title{Converts seconds to a formatted time, such as "16:26.731" (Vectorized).}
\usage{
toTime(prmSeconds)
}
\arguments{
\item{prmSeconds}{The amount of seconds to convert (ex: 986.731).}
}
\value{
The value of prmSeconds in minutes and seconds.
}
\description{
Converts seconds to a formatted time, such as "16:26.731" (Vectorized).
}
