% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inSeconds}
\alias{inSeconds}
\title{Converts a formatted time to its value in seconds (Vectorized).}
\usage{
inSeconds(raceTimes)
}
\arguments{
\item{raceTimes}{The formatted times (ex: "16:26") to convert}
}
\value{
The seconds value for each time
}
\description{
Converts a formatted time to its value in seconds (Vectorized).
}
